+++ src/components/ExhibitionSlider/components/ExhibitionInfo/index.tsx
