+++ src/components/ExhibitionSlider/index.tsx
@@ -6,7 +6,7 @@ import type { SwiperOptions } from 'swiper/types';
 import 'swiper/scss';
 import 'swiper/scss/pagination';
 import 'swiper/scss/navigation';
-import ExhibitionInfo from './conponents/ExhibitionInfo';
+import ExhibitionInfo from './components/ExhibitionInfo';
 import { ExhibitionData } from '@/types/ExhibitionData';
 import getImageURL from '@/utils/getImageURL';
